mame2000/drawgfx: cast pointer to uintptr_t for alignment checks instead of long
The 8bpp / NtoN / 8toN blockmove implementations in src/drawgfx.c
do longword-alignment checks of the form:

   while (((long)srcdata & 3) && dstdata < end)   /* longword align */
   ...
   if ((long)dstdata & 3)

casting the source/destination pointer to `long` before the AND.
The intent is portable: peel scalar pixels off the head of the run
until the inner pointer hits a 4-byte boundary, then process four
pixels at a time.

The cast type is wrong on platforms where sizeof(long) != sizeof
(void *).  On 64-bit Windows (LLP64) `long` is 32 bits while
pointers are 64 bits, so casting truncates the top half before the
AND -- in practice always producing the same low-3-bit result so
the alignment check still works, but gcc rightfully emits
-Wpointer-to-int-cast for the lossy conversion:

   src/drawgfx.c:200:25: warning: cast from pointer to integer of
       different size [-Wpointer-to-int-cast]
   src/drawgfx.c:212:21: warning: cast from pointer to integer of
       different size [-Wpointer-to-int-cast]
   ... 13 more, one per blockmove specialisation (one for each
   8bpp variant produced by the function##8 / function##16 macro
   expansions of DECLARE for NtoN_transpen_noremap{_flipx}? and
   the 8toN_trans{pen,mask}{_raw,_pri,_pri_flipx,_flipx,}? family
   in lines 2545-3333).

The correct integer type for pointer-bit arithmetic is uintptr_t
from <stdint.h>, guaranteed to be the same width as a pointer on
any conforming target.  Replace all 15 (long)pointer casts in
this file with (uintptr_t)pointer casts:

   while (((uintptr_t)srcdata & 3) && dstdata < end)
   ...
   if ((uintptr_t)dstdata & 3)

stdint.h is already reachable through driver.h (osd_cpu.h ->
<stdint.h>), so no new include is needed.

This is a portability / warning-hygiene fix only.  Generated code
is identical on platforms where sizeof(long) == sizeof(void *)
(all 64-bit Unix-likes including the libretro Linux/Mac/Android
ARM and Pi targets).  On 64-bit Windows (MINGW64 / MSVC) the cast
now uses the full pointer width before masking; the low 3 bits
checked by the AND are unchanged but the warning is gone.

TMNT2 attract-mode FNV-1a framebuffer + audio hashes unchanged
from the harness reference:
   fb hash    = 0xc11f3554bce00c6c
   audio hash = 0x7dffa26ebb0a2901
Built mame2000_libretro.so: 16391120 bytes (unchanged).  EOL
preserved (CRLF for src/*.c).

+++ src/drawgfx.c
@@ -197,7 +197,7 @@ static INLINE void blockmove_NtoN_transpen_noremap8(
 	while (srcheight)
 	{
 		end = dstdata + srcwidth;
-		while (((long)srcdata & 3) && dstdata < end)	/* longword align */
+		while (((uintptr_t)srcdata & 3) && dstdata < end)	/* longword align */
 		{
 			int col;
 
@@ -209,7 +209,7 @@ static INLINE void blockmove_NtoN_transpen_noremap8(
 
 #ifdef ALIGN_INTS /* GSL 980108 read/write nonaligned dword routine for ARM processor etc */
 
-		if ((long)dstdata & 3)
+		if ((uintptr_t)dstdata & 3)
 		{
 			while (dstdata <= end - 4)
 			{
@@ -324,7 +324,7 @@ static INLINE void blockmove_NtoN_transpen_noremap_flipx8(
 	while (srcheight)
 	{
 		end = dstdata + srcwidth;
-		while (((long)srcdata & 3) && dstdata < end)	/* longword align */
+		while (((uintptr_t)srcdata & 3) && dstdata < end)	/* longword align */
 		{
 			int col;
 
@@ -2542,7 +2542,7 @@ DECLARE(blockmove_8toN_transpen,(
 	while (srcheight)
 	{
 		end = dstdata + srcwidth;
-		while (((long)srcdata & 3) && dstdata < end)	/* longword align */
+		while (((uintptr_t)srcdata & 3) && dstdata < end)	/* longword align */
 		{
 			int col;
 
@@ -2602,7 +2602,7 @@ DECLARE(blockmove_8toN_transpen_flipx,(
 	while (srcheight)
 	{
 		end = dstdata + srcwidth;
-		while (((long)srcdata & 3) && dstdata < end)	/* longword align */
+		while (((uintptr_t)srcdata & 3) && dstdata < end)	/* longword align */
 		{
 			int col;
 
@@ -2664,7 +2664,7 @@ DECLARE(blockmove_8toN_transpen_pri,(
 	while (srcheight)
 	{
 		end = dstdata + srcwidth;
-		while (((long)srcdata & 3) && dstdata < end)	/* longword align */
+		while (((uintptr_t)srcdata & 3) && dstdata < end)	/* longword align */
 		{
 			int col;
 
@@ -2760,7 +2760,7 @@ DECLARE(blockmove_8toN_transpen_pri_flipx,(
 	while (srcheight)
 	{
 		end = dstdata + srcwidth;
-		while (((long)srcdata & 3) && dstdata < end)	/* longword align */
+		while (((uintptr_t)srcdata & 3) && dstdata < end)	/* longword align */
 		{
 			int col;
 
@@ -2854,7 +2854,7 @@ DECLARE(blockmove_8toN_transpen_raw,(
 	while (srcheight)
 	{
 		end = dstdata + srcwidth;
-		while (((long)srcdata & 3) && dstdata < end)	/* longword align */
+		while (((uintptr_t)srcdata & 3) && dstdata < end)	/* longword align */
 		{
 			int col;
 
@@ -2914,7 +2914,7 @@ DECLARE(blockmove_8toN_transpen_raw_flipx,(
 	while (srcheight)
 	{
 		end = dstdata + srcwidth;
-		while (((long)srcdata & 3) && dstdata < end)	/* longword align */
+		while (((uintptr_t)srcdata & 3) && dstdata < end)	/* longword align */
 		{
 			int col;
 
@@ -2973,7 +2973,7 @@ DECLARE(blockmove_8toN_transmask,(
 	while (srcheight)
 	{
 		end = dstdata + srcwidth;
-		while (((long)srcdata & 3) && dstdata < end)	/* longword align */
+		while (((uintptr_t)srcdata & 3) && dstdata < end)	/* longword align */
 		{
 			int col;
 
@@ -3030,7 +3030,7 @@ DECLARE(blockmove_8toN_transmask_flipx,(
 	while (srcheight)
 	{
 		end = dstdata + srcwidth;
-		while (((long)srcdata & 3) && dstdata < end)	/* longword align */
+		while (((uintptr_t)srcdata & 3) && dstdata < end)	/* longword align */
 		{
 			int col;
 
@@ -3089,7 +3089,7 @@ DECLARE(blockmove_8toN_transmask_pri,(
 	while (srcheight)
 	{
 		end = dstdata + srcwidth;
-		while (((long)srcdata & 3) && dstdata < end)	/* longword align */
+		while (((uintptr_t)srcdata & 3) && dstdata < end)	/* longword align */
 		{
 			int col;
 
@@ -3182,7 +3182,7 @@ DECLARE(blockmove_8toN_transmask_pri_flipx,(
 	while (srcheight)
 	{
 		end = dstdata + srcwidth;
-		while (((long)srcdata & 3) && dstdata < end)	/* longword align */
+		while (((uintptr_t)srcdata & 3) && dstdata < end)	/* longword align */
 		{
 			int col;
 
@@ -3273,7 +3273,7 @@ DECLARE(blockmove_8toN_transmask_raw,(
 	while (srcheight)
 	{
 		end = dstdata + srcwidth;
-		while (((long)srcdata & 3) && dstdata < end)	/* longword align */
+		while (((uintptr_t)srcdata & 3) && dstdata < end)	/* longword align */
 		{
 			int col;
 
@@ -3330,7 +3330,7 @@ DECLARE(blockmove_8toN_transmask_raw_flipx,(
 	while (srcheight)
 	{
 		end = dstdata + srcwidth;
-		while (((long)srcdata & 3) && dstdata < end)	/* longword align */
+		while (((uintptr_t)srcdata & 3) && dstdata < end)	/* longword align */
 		{
 			int col;
